requirements.txt
streamlit
opencv-python-headless
numpy
rembg
pillow
onnxruntime
